+++ src/backend/core/tests/test_utils.py
@@ -8,7 +8,18 @@ from django.core.cache import cache
 import pycrdt
 import pytest
 
-from core import factories, utils
+from core import factories
+from core.utils.dicts import get_value_by_pattern
+from core.utils.paths import get_ancestor_to_descendants_map
+from core.utils.users import (
+    get_users_sharing_documents_with_cache_key,
+    users_sharing_documents_with,
+)
+from core.utils.yjs import (
+    base64_yjs_to_text,
+    base64_yjs_to_xml,
+    extract_attachments,
+)
 
 pytestmark = pytest.mark.django_db
 
@@ -34,12 +45,12 @@ TEST_BASE64_STRING = (
 
 def test_utils_base64_yjs_to_text():
     """Test extract text from saved yjs document"""
-    assert utils.base64_yjs_to_text(TEST_BASE64_STRING) == "Hello w or ld"
+    assert base64_yjs_to_text(TEST_BASE64_STRING) == "Hello w or ld"
 
 
 def test_utils_base64_yjs_to_xml():
     """Test extract xml from saved yjs document"""
-    content = utils.base64_yjs_to_xml(TEST_BASE64_STRING)
+    content = base64_yjs_to_xml(TEST_BASE64_STRING)
     assert (
         '<heading textAlignment="left" level="1"><italic>Hello</italic></heading>'
         in content
@@ -79,13 +90,13 @@ def test_utils_extract_attachments():
     update = ydoc.get_update()
     base64_string = base64.b64encode(update).decode("utf-8")
     # image_key2 is missing the "/media/" part and shouldn't get extracted
-    assert utils.extract_attachments(base64_string) == [image_key1, image_key3]
+    assert extract_attachments(base64_string) == [image_key1, image_key3]
 
 
 def test_utils_get_ancestor_to_descendants_map_single_path():
     """Test ancestor mapping of a single path."""
     paths = ["000100020005"]
-    result = utils.get_ancestor_to_descendants_map(paths, steplen=4)
+    result = get_ancestor_to_descendants_map(paths, steplen=4)
 
     assert result == {
         "0001": {"000100020005"},
@@ -97,7 +108,7 @@ def test_utils_get_ancestor_to_descendants_map_single_path():
 def test_utils_get_ancestor_to_descendants_map_multiple_paths():
     """Test ancestor mapping of multiple paths with shared prefixes."""
     paths = ["000100020005", "00010003"]
-    result = utils.get_ancestor_to_descendants_map(paths, steplen=4)
+    result = get_ancestor_to_descendants_map(paths, steplen=4)
 
     assert result == {
         "0001": {"000100020005", "00010003"},
@@ -119,10 +130,10 @@ def test_utils_users_sharing_documents_with_cache_miss():
     factories.UserDocumentAccessFactory(user=user2, document=doc1)
     factories.UserDocumentAccessFactory(user=user3, document=doc2)
 
-    cache_key = utils.get_users_sharing_documents_with_cache_key(user1)
+    cache_key = get_users_sharing_documents_with_cache_key(user1)
     cache.delete(cache_key)
 
-    result = utils.users_sharing_documents_with(user1)
+    result = users_sharing_documents_with(user1)
 
     assert user2.id in result
 
@@ -139,12 +150,12 @@ def test_utils_users_sharing_documents_with_cache_hit():
     factories.UserDocumentAccessFactory(user=user1, document=doc1)
     factories.UserDocumentAccessFactory(user=user2, document=doc1)
 
-    cache_key = utils.get_users_sharing_documents_with_cache_key(user1)
+    cache_key = get_users_sharing_documents_with_cache_key(user1)
 
     test_cached_data = {user2.id: "2025-02-10"}
     cache.set(cache_key, test_cached_data, 86400)
 
-    result = utils.users_sharing_documents_with(user1)
+    result = users_sharing_documents_with(user1)
     assert result == test_cached_data
 
 
@@ -156,7 +167,7 @@ def test_utils_users_sharing_documents_with_cache_invalidation_on_create():
     doc1 = factories.DocumentFactory()
 
     # Pre-populate cache
-    cache_key = utils.get_users_sharing_documents_with_cache_key(user1)
+    cache_key = get_users_sharing_documents_with_cache_key(user1)
     cache.set(cache_key, {}, 86400)
 
     # Verify cache exists
@@ -182,7 +193,7 @@ def test_utils_users_sharing_documents_with_cache_invalidation_on_delete():
 
     doc_access = factories.UserDocumentAccessFactory(user=user1, document=doc1)
 
-    cache_key = utils.get_users_sharing_documents_with_cache_key(user1)
+    cache_key = get_users_sharing_documents_with_cache_key(user1)
     cache.set(cache_key, {user2.id: "2025-02-10"}, 86400)
 
     assert cache.get(cache_key) is not None
@@ -196,10 +207,10 @@ def test_utils_users_sharing_documents_with_empty_result():
     """Test when user is not sharing any documents."""
     user1 = factories.UserFactory()
 
-    cache_key = utils.get_users_sharing_documents_with_cache_key(user1)
+    cache_key = get_users_sharing_documents_with_cache_key(user1)
     cache.delete(cache_key)
 
-    result = utils.users_sharing_documents_with(user1)
+    result = users_sharing_documents_with(user1)
 
     assert result == {}
 
@@ -210,7 +221,7 @@ def test_utils_users_sharing_documents_with_empty_result():
 def test_utils_get_value_by_pattern_matching_key():
     """Test extracting value from a dictionary with a matching key pattern."""
     data = {"title.extension": "Bonjour", "id": 1, "content": "test"}
-    result = utils.get_value_by_pattern(data, r"^title\.")
+    result = get_value_by_pattern(data, r"^title\.")
 
     assert set(result) == {"Bonjour"}
 
@@ -218,7 +229,7 @@ def test_utils_get_value_by_pattern_matching_key():
 def test_utils_get_value_by_pattern_multiple_matches():
     """Test that all matching keys are returned."""
     data = {"title.extension_1": "Bonjour", "title.extension_2": "Hello", "id": 1}
-    result = utils.get_value_by_pattern(data, r"^title\.")
+    result = get_value_by_pattern(data, r"^title\.")
 
     assert set(result) == {
         "Bonjour",
@@ -229,7 +240,7 @@ def test_utils_get_value_by_pattern_multiple_matches():
 def test_utils_get_value_by_pattern_multiple_extensions():
     """Test that all matching keys are returned."""
     data = {"title.extension_1.extension_2": "Bonjour", "id": 1}
-    result = utils.get_value_by_pattern(data, r"^title\.")
+    result = get_value_by_pattern(data, r"^title\.")
 
     assert set(result) == {"Bonjour"}
 
@@ -237,6 +248,6 @@ def test_utils_get_value_by_pattern_multiple_extensions():
 def test_utils_get_value_by_pattern_no_match():
     """Test that empty list is returned when no key matches the pattern."""
     data = {"name": "Test", "id": 1}
-    result = utils.get_value_by_pattern(data, r"^title\.")
+    result = get_value_by_pattern(data, r"^title\.")
 
     assert result == []
