+++ src/backend/core/tests/test_utils_users_sharing_documents_with.py
@@ -4,7 +4,8 @@ from django.utils import timezone
 
 import pytest
 
-from core import factories, utils
+from core import factories
+from core.utils.users import users_sharing_documents_with
 
 pytestmark = pytest.mark.django_db
 
@@ -54,7 +55,7 @@ def test_utils_users_sharing_documents_with():
     doc_3_pierre_2.created_at = yesterday
     doc_3_pierre_2.save()
 
-    shared_map = utils.users_sharing_documents_with(user)
+    shared_map = users_sharing_documents_with(user)
 
     assert shared_map == {
         pierre_1.id: last_week,
