+++ src/backend/core/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Core utilities package."""
